refactor: use int for statusCode instead of uint16
To comply with the status code data type in http package,
use int instead of uint16 to avoid further type conversion.

+++ v2/dtos/responses/device.go
@@ -18,11 +18,11 @@ type DeviceResponse struct {
 	Device              dtos.Device `json:"device"`
 }
 
-func NewDeviceResponseNoMessage(requestId string, statusCode uint16, device dtos.Device) DeviceResponse {
+func NewDeviceResponseNoMessage(requestId string, statusCode int, device dtos.Device) DeviceResponse {
 	return NewDeviceResponse(requestId, "", statusCode, device)
 }
 
-func NewDeviceResponse(requestId string, message string, statusCode uint16, device dtos.Device) DeviceResponse {
+func NewDeviceResponse(requestId string, message string, statusCode int, device dtos.Device) DeviceResponse {
 	return DeviceResponse{
 		BaseResponse: common.NewBaseResponse(requestId, message, statusCode),
 		Device:       device,
