+++ v2/dtos/responses/deviceservice.go
@@ -18,11 +18,11 @@ type DeviceServiceResponse struct {
 	Service             dtos.DeviceService `json:"service"`
 }
 
-func NewDeviceServiceResponseNoMessage(requestId string, statusCode uint16, deviceService dtos.DeviceService) DeviceServiceResponse {
+func NewDeviceServiceResponseNoMessage(requestId string, statusCode int, deviceService dtos.DeviceService) DeviceServiceResponse {
 	return NewDeviceServiceResponse(requestId, "", statusCode, deviceService)
 }
 
-func NewDeviceServiceResponse(requestId string, message string, statusCode uint16, deviceService dtos.DeviceService) DeviceServiceResponse {
+func NewDeviceServiceResponse(requestId string, message string, statusCode int, deviceService dtos.DeviceService) DeviceServiceResponse {
 	return DeviceServiceResponse{
 		BaseResponse: common.NewBaseResponse(requestId, message, statusCode),
 		Service:      deviceService,
